chore: restructure `api` submodule
